refactor: update story and enhanch component file structure for gender id -tckt-365
