+++ packages/design/src/Form/components/GenderId/GenderId.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { type Meta, type StoryObj } from '@storybook/react';
-import { GenderIdPattern } from './GenderId.js';
+import GenderIdPattern from './index.js';
 
 const meta: Meta<typeof GenderIdPattern> = {
   title: 'patterns/GenderIdPattern',
@@ -24,67 +24,41 @@ const meta: Meta<typeof GenderIdPattern> = {
 
 export default meta;
 
+const defaultArgs = {
+  genderId: 'gender-identity',
+  label: 'Gender identity',
+  hint: 'For example, man, woman, non-binary',
+  required: true,
+  preferNotToAnswerText: 'Prefer not to share my gender identity',
+};
+
 export const Default: StoryObj<typeof GenderIdPattern> = {
-  args: {
-    genderId: 'gender-identity',
-    label: 'Gender identity',
-    hint: 'For example, man, woman, non-binary',
-    required: true,
-    preferNotToAnswerText: 'Prefer not to share my gender identity',
-  },
+  args: { ...defaultArgs },
 };
 
 export const Optional: StoryObj<typeof GenderIdPattern> = {
-  args: {
-    genderId: 'gender-identity',
-    label: 'Gender identity',
-    hint: 'For example, man, woman, non-binary',
-    required: false,
-    preferNotToAnswerText: 'Prefer not to share my gender identity',
-  },
+  args: { ...defaultArgs, required: false },
 };
 
 export const WithError: StoryObj<typeof GenderIdPattern> = {
   args: {
-    genderId: 'gender-identity',
+    ...defaultArgs,
     label: 'Gender identity with error',
-    hint: 'For example, man, woman, non-binary',
-    required: true,
     error: {
       type: 'custom',
       message: 'This field has an error',
     },
-    preferNotToAnswerText: 'Prefer not to share my gender identity',
   },
 };
 
 export const WithHint: StoryObj<typeof GenderIdPattern> = {
-  args: {
-    genderId: 'gender-identity',
-    label: 'Gender identity',
-    hint: 'For example, man, woman, non-binary',
-    required: true,
-    preferNotToAnswerText: 'Prefer not to share my gender identity',
-  },
+  args: { ...defaultArgs },
 };
 
 export const WithCheckboxChecked: StoryObj<typeof GenderIdPattern> = {
-  args: {
-    genderId: 'gender-identity',
-    label: 'Gender identity',
-    hint: 'For example, man, woman, non-binary',
-    required: true,
-    preferNotToAnswerText: 'Prefer not to share my gender identity',
-    preferNotToAnswerChecked: true,
-  },
+  args: { ...defaultArgs, preferNotToAnswerChecked: true },
 };
 
 export const WithoutCheckbox: StoryObj<typeof GenderIdPattern> = {
-  args: {
-    genderId: 'gender-identity',
-    label: 'Gender identity',
-    hint: 'For example, man, woman, non-binary',
-    required: true,
-    preferNotToAnswerText: undefined,
-  },
+  args: { ...defaultArgs, preferNotToAnswerText: undefined },
 };
